+++ apps/docs/src/app/04-components/_components/wireframe/wireframes/ComboBox.tsx
@@ -1,0 +1,26 @@
+"use client";
+import type { FC } from "react";
+import {
+  WFrame,
+  WIcon,
+  WInput,
+  WOverlay,
+  WText,
+} from "@/app/04-components/_components/wireframe/primitives";
+import { IconChevronDown } from "@mittwald/flow-react-components";
+
+export const ComboBoxWireframe: FC = () => (
+  <WFrame flexDirection="column">
+    <WInput justifyContent="flex-end">
+      <WIcon>
+        <IconChevronDown />
+      </WIcon>
+    </WInput>
+    <WOverlay>
+      <WText width="80%" />
+      <WText />
+    </WOverlay>
+  </WFrame>
+);
+
+export default ComboBoxWireframe;