+++ apps/docs/src/content/04-components/form-controls/password-creation-field/examples/buttons.tsx
@@ -2,8 +2,8 @@ import {
   Button,
   Label,
   IconSshKey,
+  PasswordCreationField,
 } from "@mittwald/flow-react-components";
-import { PasswordCreationField } from "@mittwald/flow-react-components/password-tools";
 import { useState } from "react";
 export default () => {
   const [password, setPassword] = useState("");