+++ packages/components/src/index/internal.ts
@@ -1,3 +1,4 @@
+export * from "@/components/ComponentUsageProvider";
 export { ControlledNotification } from "@/components/NotificationProvider/ControlledNotification";
 export { isFlowComponentName } from "@/lib/propsContext/isFlowComponentName";
 export * from "@/lib/componentFactory/flowComponent";