+++ packages/remote-react-renderer/src/RemoteRenderer.tsx
@@ -5,6 +5,10 @@ import type { RemoteReadyEvent } from "@mittwald/flow-remote-core";
 import React, { type FC } from "react";
 
 export type { RemoteRendererBrowserProps } from "@/RemoteRendererBrowser";
+export type {
+  ComponentUsageEvent,
+  ComponentUsageHandler,
+} from "@/lib/componentUsage";
 export type { RemoteReadyEvent };
 
 const RemoteRendererBrowser = React.lazy(